requirements.txt
APScheduler==3.7.0
bs4==0.0.1
Django==3.2.7
django-heroku==0.0.0
django-tinymce==3.3.0
gunicorn==20.1.0
Pillow==8.3.2
psycopg2-binary==2.9.1
requests==2.26.0
whitenoise==5.3.0
